// client/src/lib/constants.ts
export enum PropertyTypeEnum {
  // Existing
  Rooms = "Rooms",
  Tinyhouse = "Tinyhouse",
  Apartment = "Apartment",
  Villa = "Villa",
  Townhouse = "Townhouse",
  Cottage = "Cottage",

  Studio = "Studio",
  Loft = "Loft",
  Condo = "Condo",
  Bungalow = "Bungalow",
  Cabin = "Cabin",
  Chalet = "Chalet",
  Farmhouse = "Farmhouse",
  Houseboat = "Houseboat",
  RV = "RV",
  Camper = "Camper",
  MobileHome = "MobileHome",
  Duplex = "Duplex",
  Triplex = "Triplex",
  Penthouse = "Penthouse",
  Mansion = "Mansion",
  Castle = "Castle",
  SharedHouse = "SharedHouse",
  CoLiving = "CoLiving",
  StudentHousing = "StudentHousing",
  SeniorHousing = "SeniorHousing",
  ServicedApartment = "ServicedApartment",
  BoutiqueHotel = "BoutiqueHotel",

  OfficeSpace = "OfficeSpace",
  CoworkingSpace = "CoworkingSpace",
  RetailStore = "RetailStore",
  RestaurantSpace = "RestaurantSpace",
  Warehouse = "Warehouse",
  MedicalOffice = "MedicalOffice",
  IndustrialFlex = "IndustrialFlex",
  DataCenter = "DataCenter",
  ColdStorage = "ColdStorage",
  Showroom = "Showroom",
  GasStation = "GasStation",
  SelfStorage = "SelfStorage",
  ParkingLot = "ParkingLot",
  Hotel = "Hotel",
  Billboard = "Billboard",
  CellTower = "CellTower",
  AgriculturalLand = "AgriculturalLand",
  DarkKitchen = "DarkKitchen",
  LabSpace = "LabSpace",
  Cleanroom = "Cleanroom",
}